requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
deepspeed==0.16.2
diffusers==0.25.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.0.0
gradio_client==1.4.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
icecream==2.1.4
identify==2.6.9
idna==3.10
image-quality==1.2.7
imageio==2.37.0
importlib_metadata==8.0.0
inflect==7.3.1
iniconfig==2.1.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
libsvm==3.23.0.4
lightning-utilities==0.14.2
llvmlite==0.44.0
lmdb==1.6.2
loguru==0.7.3
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mmengine==0.10.6
modelscope==1.25.0
more-itertools==10.3.0
mpi4py_mpich==3.1.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
natsort==8.4.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peft==0.15.2
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyiqa==0.1.4
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.11.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
huggingface-hub==0.25.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
tb-nightly==2.20.0a20250319
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.0
torchaudio==2.5.0
torchmetrics==1.7.0
torchvision==0.20.0
tqdm==4.67.1
transformers==4.45.1
transformers-stream-generator==0.0.5
triton==3.1.0
typeguard==4.3.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
wandb==0.19.11
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
xformers==0.0.28.post2
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.19.2
